requirements.txt
keyboard==0.13.5
Pillow==9.3.0
plyer==2.1.0
pywin32==305
pyinstaller==5.7.0
